= SipStack =  * doxygen 을 위한 주석을 추가함

+++ SipStack/MonitorString.cpp
@@ -29,18 +29,34 @@ CMonitorString::~CMonitorString()
 {
 }
 
+/**
+ * @ingroup SipStack
+ * @brief 컬럼 문자열을 저장한다.
+ * @param pszValue 문자열
+ */
 void CMonitorString::AddCol( const char * pszValue )
 {
 	m_strBuf.append( pszValue );
 	m_strBuf.append( MR_COL_SEP );
 }
 
+/**
+ * @ingroup SipStack
+ * @brief 컬럼 문자열을 저장한다.
+ * @param strValue 문자열
+ */
 void CMonitorString::AddCol( const std::string & strValue )
 {
 	m_strBuf.append( strValue );
 	m_strBuf.append( MR_COL_SEP );
 }
 
+/**
+ * @ingroup SipStack
+ * @brief 컬럼 문자열에 IP 주소와 포트 번호를 저장한다.
+ * @param strIp IP 주소
+ * @param iPort 포트 번호
+ */
 void CMonitorString::AddCol( const std::string & strIp, int iPort )
 {
 	m_strBuf.append( strIp );
@@ -48,6 +64,11 @@ void CMonitorString::AddCol( const std::string & strIp, int iPort )
 	AddCol( iPort );
 }
 
+/**
+ * @ingroup SipStack
+ * @brief 컬럼 문자열에 숫자를 저장한다.
+ * @param iValue 숫자
+ */
 void CMonitorString::AddCol( int iValue )
 {
 	char	szValue[21];
@@ -56,6 +77,11 @@ void CMonitorString::AddCol( int iValue )
 	AddCol( szValue );
 }
 
+/**
+ * @ingroup SipStack
+ * @brief 컬럼 문자열에 시간 문자열을 저장한다.
+ * @param iTime 시간
+ */
 void CMonitorString::AddCol( time_t iTime )
 {
 	char	szTime[21];
@@ -64,18 +90,34 @@ void CMonitorString::AddCol( time_t iTime )
 	AddCol( szTime );
 }
 
+/**
+ * @ingroup SipStack
+ * @brief 줄의 마지막 문자열을 저장한다.
+ * @param pszValue 문자열
+ */
 void CMonitorString::AddRow( const char * pszValue )
 {
 	m_strBuf.append( pszValue );
 	m_strBuf.append( MR_ROW_SEP );
 }
 
+/**
+ * @ingroup SipStack
+ * @brief 줄의 마지막 문자열을 저장한다.
+ * @param strValue 문자열
+ */
 void CMonitorString::AddRow( const std::string & strValue )
 {
 	m_strBuf.append( strValue );
 	m_strBuf.append( MR_ROW_SEP );
 }
 
+/**
+ * @ingroup SipStack
+ * @brief 줄의 마지막 문자열에 IP 주소와 포트 번호를 저장한다.
+ * @param strIp IP 주소
+ * @param iPort 포트 번호
+ */
 void CMonitorString::AddRow( const std::string & strIp, int iPort )
 {
 	m_strBuf.append( strIp );
@@ -83,6 +125,11 @@ void CMonitorString::AddRow( const std::string & strIp, int iPort )
 	AddRow( iPort );
 }
 
+/**
+ * @ingroup SipStack
+ * @brief 줄의 마지막 문자열에 숫자를 저장한다.
+ * @param iValue 숫자
+ */
 void CMonitorString::AddRow( int iValue )
 {
 	char	szValue[21];
@@ -91,6 +138,11 @@ void CMonitorString::AddRow( int iValue )
 	AddRow( szValue );
 }
 
+/**
+ * @ingroup SipStack
+ * @brief 줄의 마지막 문자열에 시간 문자열을 저장한다.
+ * @param iTime 시간
+ */
 void CMonitorString::AddRow( time_t iTime )
 {
 	char	szTime[21];
@@ -99,17 +151,31 @@ void CMonitorString::AddRow( time_t iTime )
 	AddRow( szTime );
 }
 
+/**
+ * @ingroup SipStack
+ * @brief 저장된 문자열을 초기화시킨다.
+ */
 void CMonitorString::Clear( )
 {
 	m_strBuf.clear();
 }
 
+/**
+ * @ingroup SipStack
+ * @brief 저장된 문자열을 리턴한다.
+ * @returns 저장된 문자열을 리턴한다.
+ */
 const char * CMonitorString::GetString()
 {
 	return m_strBuf.c_str();
 }
 
+/**
+ * @ingroup SipStack
+ * @brief 저장된 문자열의 길이를 리턴한다.
+ * @returns 저장된 문자열의 길이를 리턴한다.
+ */
 int CMonitorString::GetLength()
 {
-	return m_strBuf.length();
+	return (int)m_strBuf.length();
 }
